Fix the two CI failures the first commit introduced
Both were self-inflicted by that commit, and both are the kind the local
gate would have caught if I had run it in full before pushing.

- shellcheck rejected lib_test.sh at info level: the stub `curl`/`sleep`
  are invoked indirectly by lib.sh (SC2317), the sourced lib.sh is not an
  input to that shellcheck run (SC1091), and the single-quoted `$(id)` in
  the rejection cases is deliberately literal (SC2016). Directives added,
  each with the reason.

- The file-size guard: ports.rs reached 345 lines against a 300 limit,
  because the fix's explanatory comments landed in an already-large file.

  Split into ports/mod.rs (154) and ports/tests.rs (196) rather than
  trimming the comments or adding a transition exception. The tests are
  longer than the code they cover, which is the right ratio for input
  validation on a security tool and the wrong reason to shorten either.
  Directory-module layout matches scan/probes and arp/platform.

No behaviour change; 147 tests still pass.
